requirements.txt
Flask==3.0.0
google-generativeai==0.7.0
# firebase-admin==6.5.0
# Flask-Cors==4.0.1
# gunicorn==22.0.0

